Created lzma_filters_to_str and lzma_str_to_filters for liblzma
lzma_filters_to_str converts a filter chain to string
for saving a used filter chain for a later run or debugging.

lzma_str_to_filters parses a string and creates a filter chain.
The format is {filter name}={option name}:{option value},{option name}:{option value}+
The '=' is only needed if you want to specify option value pairs.
The '+' is a delimiter between filters and is not needed at the
end of the string.
The ':' delimits option name to option value
The ',' delmites option name value pairs.